Added animations to Player's movements

+++ CS202FinalProject/CS202FinalProject/Animation.h
@@ -4,6 +4,7 @@
 #include <SFML/System/Clock.hpp>
 
 #include <vector>
+#include <iostream>
 
 class Animation
 {
@@ -11,11 +12,11 @@ private:
 	std::vector<sf::IntRect> frames; // stores the rects covering the frames
 	float frameTime; // time for each frames, in this Game assume equal
 	int cur_frame; // index of current frame
-	float cur_frame_time; // calculate current frame's time to determine if it is time to switch frame
+	double cur_frame_time; // calculate current frame's time to determine if it is time to switch frame
 public:
 	Animation();
-	Animation(const std::vector<sf::IntRect>& frames, float duration);
+	Animation(const std::vector<sf::IntRect>& frames, double duration);
 	sf::IntRect& getCurrentFrame();
-	void update(float dt);
+	void update(double dt);
 };
